Fix @TargetClass naming convention violations in com.oracle.svm.truffle.nfi.posix
